pyproject.toml
[tool.poetry]
name = "call-analytics"
version = "0.1.0"
description = ""
authors = ["Lev <lc@lev.la>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
insanely-fast-whisper = "^0.0.15"
openai = "^1.31.0"
rich = "^13.7.1"
typer = "^0.12.3"
whisper = "^1.1.10"
transformers = "^4.41.2"
optimum = "^1.20.0"
accelerate = "^0.30.1"
replicate = "^0.26.0"
python-dotenv = "^1.0.1"
gspread = "^6.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
